WIP: Codegen for array support
This commit is a solution that works for everything but arrays and for-loops. The array's are being properly created (mostly) and I am just working on bugs here or there with arrays. For-loops have not been implemented yet.

IR code looks mostly right.
Example with simplerarray.sip:
    `a = [1, 2];` is translated to:
    ```
startarray1:                                      ; preds = %entry
  br i1 false, label %implicitarray1, label %explicitarray1

explicitarray1:                                   ; preds = %startarray1
  %0 = alloca i64, i64 3, align 8
  %1 = getelementptr i64, i64* %0, i64 0
  store i64 2, i64* %1, align 4
  %elementptr = getelementptr i64, i64* %0, i64 1
  store i64 1, i64* %elementptr, align 4
  %elementptr1 = getelementptr i64, i64* %0, i64 2
  store i64 2, i64* %elementptr1, align 4
  br label %endarray1

implicitarray1:                                   ; preds = %startarray1
  %2 = alloca i64, i64 2, align 8
  %sizeptr = getelementptr i64, i64* %2, i64 0
  store i64 2, i64* %sizeptr, align 4
  %elementptr2 = getelementptr i64, i64* %2, i64 1
  store i64 2, i64* %elementptr2, align 4
  br label %endarray1

endarray1:                                        ; preds = %implicitarray1, %explicitarray1
  %3 = ptrtoint i64* %2 to i64
  store i64 %3, i64* %a, align 4
  ```

  However, the codegen for the array access is not correct yet. This could be due to an error in the codegen for the array creation, or in the codegen for the array access. I'm not sure yet.

  In simplerarray.sip, `return #a;` is translated to:
```
startarraylength2:                                ; preds = %endarray1
  %sizeptr3 = getelementptr i64, i64* %a, i64 0
  %4 = load i64, i64* %sizeptr3, align 4
  ret i64 %4
```

Yet this currently returns:
```
$ ./simplerarray
Program output: 4440596736
```

I think the problem may be that this is a pointer and not the value stored at the pointer. As I believe this address would be situated in userspace memory, but I am not certain.